+++ samples/elaborate/tetris/view/tetris_menu_bar.rb
@@ -45,8 +45,8 @@ class Tetris
             menu_item(:check) {
               text '&Pause'
               accelerator COMMAND_KEY, :p
-              enabled <= [game, :game_over, on_read: ->(value) { value && !game.show_high_scores }]
-              enabled <= [game, :show_high_scores, on_read: ->(value) { value && !game.game_over }]
+              enabled <= [game, :game_over, on_read: ->(value) { !value && !game.show_high_scores }]
+              enabled <= [game, :show_high_scores, on_read: ->(value) { !value && !game.game_over }]
               selection <=> [game, :paused]
             }
             menu_item {